devel/lua-ffi: add new port
This is a library for calling C function and manipulating C types from
lua. It is designed to be interface compatible with the FFI library in
LuaJIT. It can parse C function declarations and struct definitions
that have been directly copied out of C header files and into lua
source as a string.

https://github.com/jmckaskill/luaffi
http://luajit.org/ext_ffi.html