x11/nwg-panel: add new port
The author have been using sway since 2019 and find it the most
comfortable working environment, but...  Have you ever missed all the
graphical bells and whistles in your panel, we used to have in tint2?
It happens to me. That's why the author decided to try to code a
GTK-based panel, including best features from my two favourites:
Waybar and tint2. Many thanks to Developers and Contributors of the
both projects!

There are 8 modules available at the moment, and the author don't plan
on many more. Basis system controls are available in the Controls
module, and whatever else you may need, there's an executor for that.

https://github.com/nwg-piotr/nwg-panel